+++ VentasService/src/main/java/com/example/VentasService/repository/DetalleBoletaRepository.java
@@ -8,24 +8,24 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DetalleBoletaRepository extends JpaRepository<DetalleBoleta, Integer> {
+public interface DetalleBoletaRepository extends JpaRepository<DetalleBoleta, Long> {
     
     // Obtener detalles de una boleta específica
-    List<DetalleBoleta> findByBoletaId(Integer boletaId);
+    List<DetalleBoleta> findByBoletaId(Long boletaId);
     
     // Obtener detalles ordenados por ID
-    List<DetalleBoleta> findByBoletaIdOrderByIdAsc(Integer boletaId);
+    List<DetalleBoleta> findByBoletaIdOrderByIdAsc(Long boletaId);
     
     // Buscar detalles por producto (inventario)
-    List<DetalleBoleta> findByInventarioId(Integer inventarioId);
+    List<DetalleBoleta> findByInventarioId(Long inventarioId);
     
     // Contar productos en una boleta
-    long countByBoletaId(Integer boletaId);
+    long countByBoletaId(Long boletaId);
     
     // Eliminar todos los detalles de una boleta
-    void deleteByBoletaId(Integer boletaId);
+    void deleteByBoletaId(Long boletaId);
     
     // Obtener total de cantidad vendida de un producto
     @Query("SELECT SUM(d.cantidad) FROM DetalleBoleta d WHERE d.inventarioId = :inventarioId")
-    Integer getTotalCantidadVendidaPorProducto(Integer inventarioId);
+    Integer getTotalCantidadVendidaPorProducto(Long inventarioId);
 }
